fix: remove redundant type and refactor multipartDownload function
