+++ src/app/workers/downloadWorker.test.ts
@@ -1,9 +1,10 @@
 import { describe, expect, vi, beforeEach, test, afterEach } from 'vitest';
-import { DownloadFilePayload, DownloadWorker } from './downloadWorker';
+import { DownloadWorker } from './downloadWorker';
 import createFileDownloadStream from 'app/drive/services/download.service/createFileDownloadStream';
 import { binaryStreamToBlob } from 'app/core/services/stream.service';
 import createMultipartFileDownloadStream from 'app/drive/services/download.service/createMultipartDownloadStream';
 import { DriveItemData } from 'app/drive/types';
+import { DownloadFilePayload } from './types/download';
 
 vi.mock('app/drive/services/download.service/createFileDownloadStream', () => ({
   default: vi.fn(),
